Made first frame and last frame not overpass each other. Made memory support long double.
Former-commit-id: f672e6689da48824ea3eb4a0dcfc63b50a30fe47

+++ mrViewer/src/gui/mrvImageInformation.cpp
@@ -392,7 +392,7 @@ static void change_gamma_cb( fltk::FloatInput* w, ImageInformation* info )
     view->redraw();
 }
 
-double ImageInformation::to_memory( double value,
+double ImageInformation::to_memory( long double value,
                                     const char*& extension )
 {
     if ( value >= 1099511627776 )
